Does the tiger fire the monkey?
Business is like a jungle. And I am like a tiger. And Dwight is like a monkey that stabs the tiger in the back with a stick. Does the tiger fire the monkey? Does the tiger transfer the monkey to another branch? Pun. There is no way of knowing what goes on inside the tiger's head. We don't have the technology.